+++ src/autopas/containers/octree/OctreeNodeWrapper.h
@@ -131,7 +131,7 @@ class OctreeNodeWrapper : public ParticleCell<Particle> {
    * Get the number of all particles stored in this cell (owned, halo and dummy).
    * @return number of particles stored in this cell (owned, halo and dummy).
    */
-  [[nodiscard]] unsigned long size() const override {
+  [[nodiscard]] size_t size() const override {
     std::lock_guard<AutoPasLock> lock(_lock);
     return _enclosedParticleCount;
   }
