+++ src/autopas/containers/verletClusterLists/VerletClusterLists.h
@@ -776,7 +776,7 @@ class VerletClusterLists : public ParticleContainerInterface<Particle>, public i
    * Get the number of all particles stored in this container (owned + halo + dummy).
    * @return number of particles stored in this container (owned + halo + dummy).
    */
-  [[nodiscard]] unsigned long size() const override {
+  [[nodiscard]] size_t size() const override {
     size_t sum = std::accumulate(_towers.begin(), _towers.end(), 0,
                                  [](size_t acc, const auto &tower) { return acc + tower.size(); });
     sum = std::accumulate(_particlesToAdd.begin(), _particlesToAdd.end(), sum,
